+++ Offers/UI/EquipmentDictionary.cs
@@ -67,17 +67,20 @@ namespace Offers.UI
 
         private void удалитьToolStripMenuItem_Click(object sender, EventArgs e)
         {
-            if (dataGridView1.CurrentRow != null)
+            if (MessageBox.Show("Вы уверены?", "Удаление", MessageBoxButtons.YesNo) == DialogResult.Yes)
             {
-                var id = Convert.ToInt32(dataGridView1.CurrentRow.Cells["ID"].Value);
-                using (UserContext db = new UserContext(Settings.constr))
+                if (dataGridView1.CurrentRow != null)
                 {
-                    var equip = db.Equipments.FirstOrDefault(x => x.ID == id);
-                    db.Equipments.Remove(equip);
-                    db.SaveChanges();
+                    var id = Convert.ToInt32(dataGridView1.CurrentRow.Cells["ID"].Value);
+                    using (UserContext db = new UserContext(Settings.constr))
+                    {
+                        var equip = db.Equipments.FirstOrDefault(x => x.ID == id);
+                        db.Equipments.Remove(equip);
+                        db.SaveChanges();
+                    }
+                    Init();
                 }
-                Init();
-            }
+            }   
         }
     }
 }
